+++ Back-End Project/Areas/Admin/Controllers/CourseController.cs
@@ -2,6 +2,7 @@
 using Back_End_Project.Contexts;
 using Back_End_Project.Models;
 using Back_End_Project.Models.ManyToMany;
+using Back_End_Project.Utilits;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -36,8 +37,7 @@ namespace Back_End_Project.Areas.Admin.Controllers
         }
         public IActionResult Create()
         {
-            ViewBag.Categories = _context.Categories.ToList();
-
+            ViewBag.Categories= _context.Categories.ToList();
             return View();
         }
 
@@ -45,10 +45,20 @@ namespace Back_End_Project.Areas.Admin.Controllers
         [ActionName("Create")]
         public async Task<IActionResult> Create(CourseViewModel courseViewModel)
         {
+            ViewBag.Categories = _context.Categories.ToList();
+            if (!courseViewModel.Image.ContentType.Contains("image"))
+            {
+                ModelState.AddModelError("Image", "File type is not image .");
+                return View();
+            }
+            if (!courseViewModel.Image.CheckFileSize(1500))
+            {
+                ModelState.AddModelError("image", "file size it is the very big.");
+                return View();
+            }
 
             string path = _webHostEnvironment.ContentRootPath + "\\wwwroot\\img\\course\\" + "eheehe-" + amount + courseViewModel.Image.FileName;
 
-            ViewBag.Categories = _context.Categories.ToList();
 
             Course course = new()
             {
@@ -97,12 +107,35 @@ namespace Back_End_Project.Areas.Admin.Controllers
             };
             await _context.SaveChangesAsync();
 
+
+
+
+            EmailHelper emailHelper = new EmailHelper();
+            List<Subscribe> subscribes = await _context.Subscribes.ToListAsync();
+            foreach (Subscribe subscribe in subscribes)
+            {
+                MailRequestViewModel mailRequestViewModel = new()
+                {
+                    ToEmail = subscribe.Email,
+                    Subject = "Course Info",
+                    Body = $"<h1>Name:{course.Name}</h1>" +
+                    $"<h2>Description:{course.Desc}</h2>" +
+                    $"<h3>Buyurun Asiman mellimin kurslarindan yararlanin</h3>",
+                    
+
+                };
+            await emailHelper.SendEmailAsync(mailRequestViewModel);
+            }
+
+
             return RedirectToAction(nameof(Index));
 
         }
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (_context.Courses.ToList().Count <= 3)
+                return BadRequest();
             Course? course = await _context.Courses.FirstOrDefaultAsync(x => x.Id == id);
             if (course == null)
                 return NotFound();
@@ -116,6 +149,8 @@ namespace Back_End_Project.Areas.Admin.Controllers
         [ActionName("Delete")]
         public async Task<IActionResult> DeleteCourse(int id)
         {
+            if (_context.Courses.ToList().Count <= 3)
+                return BadRequest();
             Course? course = await _context.Courses.FirstOrDefaultAsync(c => c.Id == id);
             if (course == null)
                 return NotFound();
@@ -140,7 +175,7 @@ namespace Back_End_Project.Areas.Admin.Controllers
             Course? course=await _context.Courses.FirstOrDefaultAsync(x => x.Id == id);
             if (course == null)
                 return NotFound();
-
+            ViewBag.Categories = await _context.Categories.ToListAsync();
             CourseViewModel courseViewModel = new()
             {
 
@@ -165,8 +200,9 @@ namespace Back_End_Project.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> Update(int id,CourseViewModel courseViewModel)
         {
+            ViewBag.Categories = await _context.Categories.ToListAsync();
 
-            Course? course= await _context.Courses.FirstOrDefaultAsync(x => x.Id == id);
+            Course? course= await _context.Courses.Include(d=>d.CategoryCourses).FirstOrDefaultAsync(x => x.Id == id);
             if (course == null)
                 return NotFound();
 
@@ -177,7 +213,17 @@ namespace Back_End_Project.Areas.Admin.Controllers
 
             if (courseViewModel.Image != null)
             {
-                if(System.IO.File.Exists(path+course.Image))
+                if (!courseViewModel.Image.ContentType.Contains("image"))
+                {
+                    ModelState.AddModelError("Image", "File type is not image .");
+                    return View();
+                }
+                if (!courseViewModel.Image.CheckFileSize(1500))
+                {
+                    ModelState.AddModelError("Image", "Faylin hecmi 1 mb-dan kicik olmalidir.");
+                    return View();
+                }
+                if (System.IO.File.Exists(path+course.Image))
                     System.IO.File.Delete(path+course.Image);
                 course.Image = "eheehe-" + amount+courseViewModel.Image.FileName;
 
@@ -186,6 +232,31 @@ namespace Back_End_Project.Areas.Admin.Controllers
                     await courseViewModel.Image.CopyToAsync(stream);
                 }
             amount += 123;
+            }
+
+            foreach (var categoryCourse1 in course.CategoryCourses)
+            {
+                var catogry=await _context.Categories.Include(b=>b.CategoryCourses).FirstOrDefaultAsync(x => x.Id == categoryCourse1.CategoryId);
+                catogry.CategoryCourses.Remove(categoryCourse1);
+
+            }
+            course.CategoryCourses = null;
+            
+            List<CategoryCourse> categoryCourses = new List<CategoryCourse>();
+
+            foreach (int ids in courseViewModel.CategoryIds)
+            {
+                CategoryCourse categoryCourse = new()
+                {
+                    CourseId = courseViewModel.Id,
+                    CategoryId = ids
+
+                };
+                categoryCourses.Add(categoryCourse);
+                Category? category = await _context.Categories.Include(x => x.CategoryCourses).FirstOrDefaultAsync(c => c.Id == ids);
+             
+                category.CategoryCourses.Add(categoryCourse);
+
             }
             course.Name = courseViewModel.Name;
             course.Assesments = courseViewModel.Assesments;
